# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Supplementary Insurance Module API
  description: >
    REST API of the supplementary-insurance capstone module: config-driven
    coverage rules, a multi-stage claim workflow with a full audit trail,
    RBAC (admin / reviewer / employee / auditor), management reporting, and a
    parent-system integration seam. Monetary amounts are whole Rial — the
    currency has no fractional unit in everyday use (ADR-0003). This document
    is the single source of truth for the wire format — the Go transport DTOs and the frontend's
    generated TypeScript types both conform to it (ADR-0002).
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    Error:
      description: Error envelope (all non-2xx responses).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ServiceType:
      type: object
      required: [id, code, name, created_at]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        created_at: { type: string, format: date-time }
    InsuranceContract:
      type: object
      required: [id, name, start_date, end_date, is_active, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CoveragePlan:
      type: object
      required: [id, contract_id, name, description, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        contract_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    CreateServiceTypeRequest:
      type: object
      required: [code, name]
      properties:
        code: { type: string, description: "Stable machine code (lowercase, digits, underscores), unique." }
        name: { type: string, description: "Display name (Persian)." }
    CreateContractRequest:
      type: object
      required: [name, start_date, end_date]
      properties:
        name: { type: string }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time }
        is_active: { type: boolean }
    CreatePlanRequest:
      type: object
      required: [contract_id, name]
      properties:
        contract_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
...
paths:
...
  /service-types:
    get:
      operationId: listServiceTypes
      responses:
        "200":
          description: All service types.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ServiceType" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createServiceType
      description: Admin only. Adds a claimable service category to the catalogue.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateServiceTypeRequest" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceType" }
        default: { $ref: "#/components/responses/Error" }
  /contracts:
    get:
      operationId: listContracts
      responses:
        "200":
          description: All insurance contracts.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InsuranceContract" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createContract
      description: Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateContractRequest" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InsuranceContract" }
        default: { $ref: "#/components/responses/Error" }
  /plans:
    get:
      operationId: listPlans
      parameters:
        - name: contract_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Coverage plans.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CoveragePlan" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createPlan
      description: Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePlanRequest" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CoveragePlan" }
        default: { $ref: "#/components/responses/Error" }
...